Fix for bug #238
Really, you have to stare at the patch for a while to fully understand how the changes I made affect anything at all. However, in a nut shell, I tweaked the limits for the song so that the initial sanity-enforcement wouldn't clip the song number, and adjusted the actual menu code. Before, it was the other way around.

That made no sense, but, again, look at the diff to see.


git-svn-id: https://rpg.hamsterrepublic.com/source/wip@822 7d344553-34f0-0310-a9b1-970ce8f1c3a2