OGG, MP3, etc. support
Ok, so, here's the story so far:

Audiere is a C++ library, so I created Audwrap to turn it into a C library.

That done, we now have full support for MP3, OGG, MOD, IT, S3M and XM file formats, in addition to WAV and MIDI (however, WAV is now played using Audiere for consistency).

So, a few caveats:

1. First and foremost, I made the arbitrary decision to not allow WAV as a music format. This is to discourage 60 meg wave files used as music. MP3 or OGG streamed will do just fine. While I'm at it, I'll mention that you can't use BAM or MIDI as a sound effect (since they're not played with Audiere, and would interfere with the regular music)

2. The music functions handle streaming vs non streaming by themselves (i.e., MP3s and OGGs always stream, nothing else does). But, for regular sound effects, I figure it's best if the user chooses whether or not they're streamed. The tracker formats cannot be streamed anyway, so it's basically a WAV, OGG and MP3 thing. However, I'll deal with that later. Right now, they use the same rules to decide whether or not to stream.

3. Right now, two new DLLs are required: audiere.dll (which is the work horse), and audwrap.dll (my wrapper). I wish dearly I could link either one statically, but, alas, I don't know how to make a static library instead of a DLL. I need to play around with the compiler settings a bit...

(also, a note about this commit itself: It includes a few non-mp3 related changes, mostly in prep for other work I'm going to be doing shortly. sorry if it breaks anything (it shouldn't))

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@875 7d344553-34f0-0310-a9b1-970ce8f1c3a2