Two sound related changes
Fixed browse() so that sound effects are really stopped when they are not being played.

Also, added a 'Special Sound Effects' menu to the general preferences menu, for choosing interface sounds (like the ones Kizul has been uploading to the wiki).

On that topic, actually adding the triggers it going to be easy, but a pain, on account of the fact that there's not centralized menu subroutine. However, I will plug away at it until it's done.


git-svn-id: https://rpg.hamsterrepublic.com/source/wip@895 7d344553-34f0-0310-a9b1-970ce8f1c3a2