Added "none" option to map and battle music, previous "none" renamed to "silence"
A map with music set to none continues after a battle to play the music that it was playing before the battle, this seemed the most logical. Victory music is also ignored. After a textbox with music with "reset music" set to off, a map with 'none' music reverts to its old tune, just like a regular map. I'm not 100% about these behaviours.

Made a small improvements to zintgrabber and added a little documentation

Renamed stopsong in allmodex.bas to pausesong (which is what it really does), and added a wrapper stopsong to GAME, which updates presentsong. changed instances of pausesong in CUSTOM and common.bas to pausesong.

Because of that change, fixed a bug where currentsong didn't return -1 when music wasn't playing - also added this fact to the dictionary.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@987 7d344553-34f0-0310-a9b1-970ce8f1c3a2