Making audwrap cross-platform (and commenting)
A Makefile should probably be generated for those on Linux, or MingW, but I don't posess the knowledge to create one. However, once that's done, then it should be easy to compile this on Linux. (I got rid of stdafx.h, stdafx.cpp, and the <windows.h> include, since they're totally unneeded)


git-svn-id: https://rpg.hamsterrepublic.com/source/wip@1061 7d344553-34f0-0310-a9b1-970ce8f1c3a2