New map resizing screen, which zooms and lets you resize from any map edge.
We now have 3 functions generating minimaps. They should be combined, but I don't know what to do in the tilemap editor: it would be nice to combine the TAB and ~ functions into something that zooms and lets you navigate the map quickly, and not just in the tileset editor. Ideas?

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@1131 7d344553-34f0-0310-a9b1-970ce8f1c3a2