Fixing a few "implicit variable allocation" warnings.
All but one were a case where a variable was implicitly assigned to in a gosub that was physically below where it was first used. To fix these, I added a shared module variable, which can be refactored later when the gosubs are subified (as if we'll remember)

However, there was one case where the variable in question really was an orphan, so I commented it out. It's in custom.bas on line 1261. This should be fixed ASAP, but I'm not familiar enough with that code to fix it myself.



git-svn-id: https://rpg.hamsterrepublic.com/source/wip@1166 7d344553-34f0-0310-a9b1-970ce8f1c3a2