Revising the windows build batch files
I've modified the batch files to be a bit more intelligent and less confusing:

Make.bat: It's main job is to ensure OHRGFX and OHRMUSIC are set to something. If you specify nothing, it will set the defaults (fb and sdl respectively) and then call makegame and makeedit. If you *do* specify something, and they/it are detected as gfx_ or music_ modules, they will be assigned to OHRGFX and OHRMUSIc. Anything further will be passed to makeedit and makegame (which will then be passed to fbc).

Further, if you want to specify parameters, but not change the modules, you can pass a tilde (~) as the first parameter, to force it to not parse anything more, and to pass everything else straight through.

Makeedit.bat and Makegame.bat: I renamed make????-gfx-music.bat to these, and changed all references to %1 and %2 to %OHRGFX% and %OHRMUSIC% respectively. That's all. The old make????.bat files have been removed.

Note that makeedit and makegame cannot override the default backends -- they will always use OHRGFX and OHRMUSIC.
 


git-svn-id: https://rpg.hamsterrepublic.com/source/wip@1579 7d344553-34f0-0310-a9b1-970ce8f1c3a2