Fixing James' implementation of the pixel dissolve
I thought you were intending to have this effect as the "Default" (i.e. old style that old games use), as evidenced by including a "Default" option in the global game settings (otherwise, what would the default be when choosing the default?). I fixed it to follow that idea.

Further, I corrected a bug in the enemy editor where a caption was not set for the new dissolve type, causing a crash. Fixed by, again, relegating the dissolve to "Default"


git-svn-id: https://rpg.hamsterrepublic.com/source/wip@1671 7d344553-34f0-0310-a9b1-970ce8f1c3a2