Make the formation editor preview backdrop animation, and default new enemies to center of field (bug 124), as well as other improvements. Also, backdrops now wrap around to number 0 when you 'animate' past the last one.
Also, I was initially going to make the formation editor treat x,y positions as the bottom center of each enemy sprite, this is a very nice improvement, however I realised that it would make spawned enemies appear out of position as placed in the editor (on the other hand, correct spawn placement currently depends on spawned enemy size). Maybe we can fix this later.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@2011 7d344553-34f0-0310-a9b1-970ce8f1c3a2