Remove fademusic (and related calls to setfmvol) This resolves bug 678 "Ogg music problems in game with werewaffle version" The loss of automatic fading is barely even noticeable in game play. I doubt anybody who doesn't read whatsnew.txt will ever realize anything has changed.
The only remaining volume manipulation code is the volume slider in the menu and
the two plotscripting commands for manipulating volume.

Note that although music_native's implementation of fading was much better, I removed
it too, since automatic music fading is a complicated and totally unecessary feature.

Also removed a bunch of music_*.bi files that were not actually used anywhere at all.


git-svn-id: https://rpg.hamsterrepublic.com/source/wip@2283 7d344553-34f0-0310-a9b1-970ce8f1c3a2