Added code to load a document tree in. Replaced the example to test this.
However, there was a tiny bug before as well, which gave "nodes with children" the wrong node type. This means that the "test.rld" created by the previous version of reloadtest is corrupted, and won't load. Since the new reloadtest doesn't create a document now, you can fix the existing one by hex editing bytes 0x0F and 0x16 to be 7, not 8.


git-svn-id: https://rpg.hamsterrepublic.com/source/wip@2440 7d344553-34f0-0310-a9b1-970ce8f1c3a2