Two things: First, I fixed the bad cross-references in the dictionary.
Second, I've added a new script which currently converts the dictionary to a wiki-friendly format. It creates a bunch of files, one for commands starting with each letter of the alphabet, and an index. This is necessary because the server chokes on the full import package. Alas, we must upload these files manually, one-by-one

In the future, the script will take care of this, and log in as the Plotscripting Dictionary Bot, and upload each file for us. This will take a bit of time to develop, however, as HTTP in PHP is a pain.


git-svn-id: https://rpg.hamsterrepublic.com/source/wip@2500 7d344553-34f0-0310-a9b1-970ce8f1c3a2