Removed the 'cmd' argument to all the bound_* and valid_* functions; instead commandname() is used. If anyone really misses the spaces in names ("free slice" instead of "freeslice"), that can be fixed. Command name also now optional in bound_arg call. And bound_arg now takes zstring ptrs, because this makes it around 40 times faster, in my tests.
Also renamed bound_* (except for bound_arg) to valid_*: the difference between bound_* and valid_* seemed to be that only valid_* would raise errors, a difference that will disappear.

Fun bonus: removing all those arguments reduces the size of Game by 11kB.

All of this a cinch with the improved fb-search-forward-by-arg. Maybe I'll write a FB mode for Emacs!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@3068 7d344553-34f0-0310-a9b1-970ce8f1c3a2