gfx_sdl (bug 619) fixes and improvements, and added post_terminate_signal so that closing the window works with all backends.
The fundamental problem, compounded by all sorts of unfortunate assumptions and irrational code, is that gfx_sdl reports numlock and capslock (but not scrolllock) key presses different from fb and alleg: it reports whether numlock/capslock are on, rather than whether the key is down. This will probably require listening to keypress events to work around, which probably means making the pollingthread gfx_fb specific (not sure about gfx_alleg)

bug 619 is basically fixed as far as I can tell, but I'm not happy yet. Much of the fix is temporary.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@3080 7d344553-34f0-0310-a9b1-970ce8f1c3a2