The new gfx_sdl backend can be compiled using the Makefile and used by the engine. Pass 'sdlpp' to makegame/makeedit. It runs, but doesn't work very well. (Not tested on linux)
Also, some attempt to compile and link it statically, but this is a long way off working.

I changed capitalisation of current gfx interface function pointer variables from gfx_*, io_* to Gfx_*, Io_* so that they don't collide with the real functions, which was confusion gdb, and prevents static linking.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@3330 7d344553-34f0-0310-a9b1-970ce8f1c3a2