Added a bunch of comments to reload.bas, hopefully to make it clear what everything does
Also, fixed a bug where I was using Callocate in CreateDocument, but delete in FreeDocument. I can't believe that was sitting in there for so long! I'm just glad that the Doc type didn't have a constructor or destructor...

See here for more detail as to why mixing allocation methods is bad: http://www.codeproject.com/KB/tips/newandmalloc.aspx


git-svn-id: https://rpg.hamsterrepublic.com/source/wip@3467 7d344553-34f0-0310-a9b1-970ce8f1c3a2