Add 'fuzzfactor' option to fuzzyrect, giving the amount of transparency as a percentage.
This naturally leads mostly to very ugly patterns, so I took some time to adjust the pattern generated for every fuzzfactor. So the fuzzfactor has greater use for specifying a texture than for finetuning the transparency (due to lack of gamma calibration... apparent transparency... yadda, yadda...)

If anyone wants to modify the magic constants a bit, go ahead, but let's settle on them soon.

Of course, having 100 largely handpicked textures is an artificial limitation, but you'll be able to paint your own with scripting eventually.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@3669 7d344553-34f0-0310-a9b1-970ce8f1c3a2