xml2reload: Specifying iso-8859-1 encoding isn't enough, you have to translate the UTF8 that libxml2 gives you back into it. Amusingly the procedure is nearly identical to decoding a Base64 string.
And it turns out that the build of libxml2 that Mike provided is missing the crucial function for this, UTF8Toisolat1.
I couldn't easily find a larger build and get everything to link. I'm horribly pissed off at this XML stuff I volunteered for, so I just copied the function from the libxml2 source and stuck it in yet another C file. It can be deleted if someone has the patience to rebuild libxml2.a

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@3683 7d344553-34f0-0310-a9b1-970ce8f1c3a2