Extend scancodes.hsi, and enable new scancodes for scripts. Added new general bitset, "Enable better scancodes for scripts".
If that bitset is not turned on, then wholy new scancodes like key:LeftAlt will still work, but ones which were previously the same, like key:Numpad2 and key:Down, will act as if both keys had the same scancode. Turned on by default in new games.

I also reformatted the list of scancodes in the dictionary, no longer providing the value of each constant, to encourage people to actually use scancode.hsi. That'll teach 'em!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@3990 7d344553-34f0-0310-a9b1-970ce8f1c3a2