Yikes! Fix hero data corruption bug in game upgrading (hero handle points fixbit) that's been there over three years.
It was very hard to notice though. TEST.RPG has been affected by this. If you go to the 2nd spell list of heroes 1 and up, and create spells in the 2nd- and 3rd- to last slots, you'll notice that the 3-rd last one defaults to "Learned at Level 23", and the 2nd last one has an invalid level learnt that causes it to display whatever's in the slot above.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@4070 7d344553-34f0-0310-a9b1-970ce8f1c3a2