Add extremely messy Mac OSX implementation of spawn_and_wait. Which isn't quite working.
If you want to display something in a console, you should use Terminal,
there's no other option aside from packaging your own terminal emulator.
If you want to get Terminal to run something, you should write an AppleScript
to send commands to Terminal. It's not a commandline program, there's no other
way to communicate with it. To run an AppleScript, we run osascript
from the commandline. To do that, you use SHELL. As far as I can tell, this
is the proper way to do things.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@4281 7d344553-34f0-0310-a9b1-970ce8f1c3a2