Quick and dirty fix for zone editor bug: it was possible to edit zones in view mode because the tools weren't disabled in all cases.
Unfortunately Pepsi insisted that this was a terribly useful feature, not a bug. So I decided to meet him half way, and allow only the draw tool (there's no room for the toolbar), and only when the current zone is locked, so that the zone won't be automatically changing as you move about.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@4282 7d344553-34f0-0310-a9b1-970ce8f1c3a2