Remove herow(); instead use (sprite component of) hero walkabout slices in the various hero pickers.
Now heroes will display the same on the map and in the pickers, yay.
Note that these menus call set_walkabout_frame on the walkabout slices, but
this shouldn't be a problem, because displayall resets them (via
update_walkabout_slices). Also, since only the sprite component is used,
visibility of the slice walkabouts doesn't matter.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@4791 7d344553-34f0-0310-a9b1-970ce8f1c3a2