Fixed a freeze (gfx_directx only) redraw problems in the slice collection editor:
Holding down left/right to browse between collections did not redraw the screen, and in fact in gfx_directx it caused the editor to freeze (this seems to be a gfx_directx bug: it does not implement io_waitprocessing() even though it apparently should). James fixed the freeze for all other backends in r4636. In general, redraws were delayed by 1 tick, and it was probably possible to crash the editor by holding down several keys at once. IMO the 'state.need_update = YES' paradigm is flawed and we should go back to using dozens of GOTOs!

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@4859 7d344553-34f0-0310-a9b1-970ce8f1c3a2