Woe! Work around SDL_EnableUNICODE possibly creating dead keys on X11
SDL throws away a key event if XFilterEvent indicates that it's part of a
compose keysequence. This is exactly what the Xlib documentation says it should
do, but SDL shouldn't, I think this is a bug. Looking through the other SDL
platform code, X11 seems to be the unique one which behaves like this. It
causes a couple dead keys on most non-US layouts. Therefore I'm forced to
selectively enable text input, with the new io_enable_textinput function, and
adding a new argument to setkeys. This argument should be YES in any loops that
call getinputtext, such as indirectly through strgrabber.

Also rewrote intgrabber to use scancodes instead of getinputtext, like it used
to be.

Of course, all our help files give the wrong key sequences for non-US keyboards
anyway, but some unknown key working is better than nothing at all.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@4979 7d344553-34f0-0310-a9b1-970ce8f1c3a2