A suite of benchmarks for HamsterSpeak.
Because I lost my original benchmark scripts years ago. These are much more
comprehensive, and the results are surprising...

This .rpg was created with Zenzizenzic. The scripts should work with WIP versions
as far back as Jan 2009, when slice commands were added, if abs is replaced with
a script. Would be interesting to see benchmark scores over time.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@5300 7d344553-34f0-0310-a9b1-970ce8f1c3a2