New "switch" syntax: do after case optional; can write "case(else)". Also disallow non-constant expressions in cases.
(cases containing non-constant expressions will continue to work in existing games)

The old syntax is still supported, in fact you can mix the two styles freely in a single switch.

But note that
  case (1)
  case (2) ...
is NOT an empty case(1), it's still equivalent to
  case (1, 2) ...

Also, an else() inside a switch is an else case only if not attached to an if.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@5360 7d344553-34f0-0310-a9b1-970ce8f1c3a2