gfx_sdl/fb smoothing did not work at 4x zoom.
This is the obvious way to extend it to 4x, though it doesn't look that smooth:
diagonals are jagged, for example. On the positive side, single pixel dots
aren't butchered like they are in 2x or 3x

Also some commented out code to do 4x by applying 2x twice, which looks pretty
arty with lots of triangles and diagonals, but really butchers some text.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@5601 7d344553-34f0-0310-a9b1-970ce8f1c3a2