Modify vector.bi to work around a FB -gen gcc bug with aliasing multiple prototypes to the same function
Also fixed a couple problems in vectortest, which was not passing since my
errorlevel changes.

Some of the v_* functions now require parentheses around there arguments,
but no existing code needed to be changed.
It would be possible to make the parentheses optional for all of them by
providing a wrapper function for each.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@5619 7d344553-34f0-0310-a9b1-970ce8f1c3a2