Fix "get hero stat cap" and "set capped hero stat"; made "get hero stat cap" a builtin for proper error checking
In r563 the wrong order of the arguments to "set capped hero stat" was
changed, but the same problem in the body of that script was not
fixed. In fact this command has never worked.

So of course autotest.hss needed recompiling (passes now).
There are several other scripts in plotscr.hsd that need to be replaced
with commands to handle an increase in number of stats

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@5700 7d344553-34f0-0310-a9b1-970ce8f1c3a2