Stop shouting type declarations, use lowercase instead except for UDTs
That includes STRING, INTEGER, UINT8, VECTOR, BOOL, AS, PTR, ANY, BYVAL and BYREF
Only changed these keywords when they appear as type declaration, not the
STRING function, AS in INSTR, OPEN, SELECT, or ANY as an initialiser.

This cleans all files except editedit.bas, editrunner.bas and reloadedit.bas
which are consistent in always using CAPS.

James previously lowercased most INTEGER and AS, in eg. r4777

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@5863 7d344553-34f0-0310-a9b1-970ce8f1c3a2