Change subscript implementation, introducing separate nonlocal variable script nodes (previous subscript impl. incompatible)
This removes the restriction that a subscript could only be called from its
parent. Now it can always be called if it's in scope.

Also, I added a limit of 100 local variables per script. (BMR had
reported that there was a problem using more than 256 anyway). While
the file format allows up to 256 (also arbitrary) I set this limit
keeping in mind bytecoding encoding in the new interpreter (7 bits)
and leaving some free variables for temporaries.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@6068 7d344553-34f0-0310-a9b1-970ce8f1c3a2