Fix gengcc=1 with FB 0.91
I replaced a call to malloc with ALLOCATE since for some reason FB was generating
two declarations of malloc (this is normal: one is for malloc, the other for
ALLOCATE), but one of them with an 'integer' (C typedef) type instead of
'uinteger'. Which is very strange because both crt/stdlib.bi and the fbc internal
function type table use unsigned arguments. (No the fix to cutil.bi didn't
fix it.)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@6320 7d344553-34f0-0310-a9b1-970ce8f1c3a2