New item screen now shows the description of the currently selected item.
expand_slice_text_insert_codes() does not descend into children marked as planks.
This is because the screen collection and the plank should be handling their text insert
codes separately, and in some cases they can reasonably conflict with one another.
For example, the item screen and the item planks will support some of the same codes,
but for a plank it means the item associated with that plank, and for the screen it
means the currently selected item (if any)

When loading a plank that does not use SL_PLANK_HOLDER, apply that lookup code to
the plnak's root, since ti is used in several places in the plank menu code to
identify the plank


git-svn-id: https://rpg.hamsterrepublic.com/source/wip@6646 7d344553-34f0-0310-a9b1-970ce8f1c3a2