Add partial joystick support to keyval by moving joystick handling from keyispressed into script_keyval.
For joystick scancodes, keyval will return 3 if the key is down,
and 0 if not.
Real support needs to be added directly to setkeys. I actually started on it
half a year go.

Technically untested, but keyval should work if and only if keyispressed does

Also documented that waitforkey(anykey) can return joystick scancodes

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@6786 7d344553-34f0-0310-a9b1-970ce8f1c3a2