Multi-hit attacks are now cancelled once all targets are dead and dissolved (unless the attack can hit dead targets)
This is accomplished by skipping ahead in the attack animation to
the closing part. Added two new attack animation commands.

The flaw is that if there are multiple targets, the dead ones continue
to be hit until they're all dead. That could be fixed.

Of course, this annoying behaviour was intentional in parody games
like ARFENHOUSE, but nearly all games will want to avoid it.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@6860 7d344553-34f0-0310-a9b1-970ce8f1c3a2