Enable the .offset member of Frame: setting this will cause frame_draw to offset where it draws
I want to use this for dissolve animations which want to draw 'outside'
the bounds of the Frame (to the left and up), by returning a larger Frame.

Possibly the commands to flip a Frame should flip the offset too;
I'm not sure. But it doesn't currently matter for dissolved Frames.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7006 7d344553-34f0-0310-a9b1-970ce8f1c3a2