Implement in-Custom options to select the window size (zoom factor). Only available in the hidden Resolution menu.
The window size defaults to 80% of the screen width (or as near as possible),
and 50% when live-previewing.

Setting it to 100% is described as Maximised, however none of the gfx backends
support this yet. It would use a maximised window with black bars around the
edges, or probably scaled to fit in the case of gfx_directx.

I decided to just get rid of the previous genDefaultScale option
(probably never used by anyone) since letting people select the exact zoom
doesn't seem worth the complexity. You can still select a zoom using
commandline options, which overrides everything including the check
whether it's too large.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7540 7d344553-34f0-0310-a9b1-970ce8f1c3a2