64bit: redefine 'integer' as 'long' when doing a 64 bit build. Also add [u]int64 types and fix bool, and add boolint.
(Recall that 'integer' is 64 bit in 64-bit FB, while 'long' is always 32 bit.
However integer and long count as different types in 32-bit FB, which is
because int and long are mangled differently in C++, and FB is C++ compatible.)

You can use the use_native_integer() macro to revert to native 64 bit integer
before including troublesome headers, and use_32bit_integer() to go back.
Headers in FB 1.04 are meant to be more or less 64-bit safe by using 32-bit
'long' instead of 'integer', but for some reason integer still appears in
places. (E.g. in SDL.bi only some iconv stuff which doesn't seem to be part
of the public SDL API uses integer.) For now, only windows.bi is protected
with these macros.

Use boolint in C/C++ headers to indicate an int which is treated as a bool,
rather than using bool (called 'boolean' in FB, new in 1.04)

Also remove some stuff that shouldn't have been in config.bi

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7608 7d344553-34f0-0310-a9b1-970ce8f1c3a2