64bit: Fix a whole lot of compile errors in 64 bit builds
It turns out that the handling of font-changing markup in printstr was
completely wrong and should have crashed.

I have changed our declarations of FB rtlib functions to match the ones in
FB 1.04, which typically use ssize_t instead of integer. So in older versions
of FB, all of these declarations are still the same as long as you do a 32-bit
build.

The only unpleasant part was that FB refused to allow passing an array of
ENUM SliceTypes to int_array_find, regardless of which overloads I added.
I guess that's a bug, since it has no trouble casting from ENUM in 32 bit.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7609 7d344553-34f0-0310-a9b1-970ce8f1c3a2