cleanuptemp did nothing useful, so I deleted it! Instead, just killdir tmpdir.
cleanuptemp was causing a "Could not rmdir: Directory not empty" message
every time you quit, because it tried to delete tmpdir before playing.tmp.
It was uselessly manually deleting everything in tmpdir in the wrong order,
recreating it, and then afterwards it gets deleted again.

If playing an .rpgdir then cleanuptemp would delete all the .tmp files lumps
in the .rpgdir, but we should never write any temp lumps there anyway.

Not to be confused with deletetemps, which deletes just tmpdir/*.tmp and is
necessary.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7662 7d344553-34f0-0310-a9b1-970ce8f1c3a2