uiOutline (black) was being used as the text background colour! This affected anywhere that a non-zero (non-transparent) text bg colour was used with non-edged text.
This was broken a couple months ago in r7587, when I failed to noticed
the old unload_font Sub was initialising outline_col to -1.
Switched to using outline_col==0 for no outline colour.

I also fixed font_create_shadowed and font_create_edged to not designate a new
outline colour if the basefont already has one.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7720 7d344553-34f0-0310-a9b1-970ce8f1c3a2