sliceedit: Editing the root slice of a collection now works properly.
I had mistakenly thought that none of its attributes were saved, but actually
only the position and size were erased.
slice_editor() used to save and restore slice position and size when using
ctrl+F to edit a slice "in full screen", but I couldn't see any reason to
save the size, even if the slice is set to fill, so I removed that.

This was one of my Callipygous TODO list items.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7752 7d344553-34f0-0310-a9b1-970ce8f1c3a2