It's no long necessary to call UpdateScreenSlice to handle window size changes.
There were dozens of menus in Custom that failed to do so, often resulting in
parts of the menu being drawn off-screen if you shrunk the window, when after
reentering the menu.

UpdateScreenSlice is now called from within allmodex, but can still be called
by menus that want to know when the size has changed.

Also, resizepage now uses uiBackground instead of 0 to pad pages.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7762 7d344553-34f0-0310-a9b1-970ce8f1c3a2