Move editbitsets() and the general, backcompat, and activetime battle bitsets into a new common_menus.bas file.
This file is shared with Game. The idea is to allow accessing the backcompat bitsets menu.

I didn't put this stuff in common.rbas because we have way, way too much in there, and
it takes reloadbasic.py way too long to process it. Heaps of stuff from common.rbas could be
moved out (or may just move out the reloadbasic stuff from common.rbas, since it's loading/saving
code that doesn't belong in there anyway).

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7856 7d344553-34f0-0310-a9b1-970ce8f1c3a2