Add "menu item at pixel" (still buggy), and a test case for it in test.rpg
It goes crazy if the menu has scrollbars, so I put a warning in plotdict.

This is a simple wrapper around the mouse support James wrote a while ago.

Made various changes to menus 0 & 1 in test.rpg. Menu items are selected
by mouse-hover; that's not meant to be highly useable, but just for testing.
You can also click on a non-selected menu to bring it forward.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7926 7d344553-34f0-0310-a9b1-970ce8f1c3a2