The spriteset browser no longer stores loaded sprites on a video page
This removes almost all remaining uses of stosprite/loadsprite in the engine!

For now I've replaces the video page with an allocate'd memory buffer of
concatenated placer() arrays. This is temporary; I'll later switch to Frames.

When reading this patch note that sprite_editor also uses vpage 3 to store
the undo history, so it was necessary to do a lot of saving/loading when
calling sprite_editor, which I removed in favour of just passing the edited
sprite + palette back in placer() + workpal()

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7940 7d344553-34f0-0310-a9b1-970ce8f1c3a2