Macro recording and replay. Leaving this undocumented for now.
This works everywhere in Custom and Game.
Press Ctrl+F11 to start recording a macro, and Ctrl+F11 to finish.
Press Ctrl+F12 to play back the recording, or Shift+Ctrl+F12 to play it back
multiple times in a row, for those repetitive tasks. I implemented this to
ease importing huge numbers of backdrops.

The temporary .ohrkeys file is stored in tmpdir and deleted when you quit.
Maybe it should go in the CWD instead (e.g. next to the .rpg file).
Ideally there would be a way to load/save/select macros, but it doesn't seem
worth the trouble. This isn't Angband.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@7990 7d344553-34f0-0310-a9b1-970ce8f1c3a2